Revert refactoring tomfoolery.
